Extract scrapers and date provider out of test classes.
